Extract the shareable parts from the virtual guests list page
In order to share code with similar pages like the virtual networks one,
extract as much as possible of the virtualization guests page into a
component
