+++ web/html/src/manager/virtualization/list.utils.js
